0.8.9

+++ project/jimmer-spring-boot-starter/src/main/java/org/babyfish/jimmer/spring/repository/support/QueryExecutors.java
@@ -92,7 +92,7 @@ public class QueryExecutors {
                     }
                     if (returnType == Page.class) {
                         if (pageable != null) {
-                            int rowCount = query.count();
+                            long rowCount = query.count();
                             List<Object> entities = query.limit(pageable.getPageSize(), (int) pageable.getOffset()).execute();
                             return new PageImpl<>(entities, pageable, rowCount);
                         }
